requirements-dev.txt
appdirs==1.4.3
attrs==19.1.0
backports.csv==1.0.7
cached-property==1.5.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
deprecation==2.0.7
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-bootstrap-form==3.4
django-import-export==1.2.0
django-outbox-base-layout==1.2
et-xmlfile==1.0.1
idna==2.8
isodate==0.6.0
jdcal==1.4.1
lxml==4.4.1
odfpy==1.4.0
openpyxl==2.6.3
packaging==19.1
Pillow==6.1.0
pycep-correios==3.2.0
pyparsing==2.4.2
python-decouple==3.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
zeep==3.4.0
